+++ CHANGELOG.md
@@ -25,6 +25,7 @@ straightforward as possible.
 - Added limit of registers of characters to 10
 - Added the funtionality of searcher, this functionality only makes macth when the search is == to the name that is searching
   because the api haven't available the searcher for some letters
+- Added the component of alerts
  
 ### Changed
 - Changed in the menu the item favorites, now it's acceded since home in the option "View all"